[6.x] Extract CallOtherCommands feature from Illuminate\Console\Command
At the moment Illuminate\Console\Command is tightly couple with Laravel
via $this->laravel->make() usage, this changes make it possible to reuse
CallOtherCommands feature available for Laravel Command to standalone
Symfony command.
